feature: 예외정의 및 ExceptionHandler 구현 (#4)
- ExceptionHandler 및 예외 정의
- 인증정보 조회 로직 수정

+++ src/main/java/oo/kr/shared/global/security/jwt/RefreshTokenService.java
@@ -1,6 +1,8 @@
 package oo.kr.shared.global.security.jwt;
 
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
+import oo.kr.shared.global.exception.type.jwt.ExpiredRefreshTokenException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -23,13 +25,16 @@ public class RefreshTokenService {
   @Transactional
   public RefreshToken findTokenInfo(String accessToken) {
     return refreshTokenRepository.findByAccessToken(accessToken)
-                                 .orElseThrow(() -> new RuntimeException("리프레시 토큰이 존재하지 않습니다."));
+                                 .orElseThrow(ExpiredRefreshTokenException::new);
   }
 
   @Transactional
   public void removeRefreshToken(String accessToken) {
-    RefreshToken refreshToken = refreshTokenRepository.findByAccessToken(accessToken)
-                                                      .orElseThrow(RuntimeException::new);
+    Optional<RefreshToken> optionalRefreshToken = refreshTokenRepository.findByAccessToken(accessToken);
+    if (optionalRefreshToken.isEmpty()) {
+      return;
+    }
+    RefreshToken refreshToken = optionalRefreshToken.get();
     refreshTokenRepository.delete(refreshToken);
   }
 }
